+++ src/admin/admin.ts
@@ -1,5 +1,5 @@
 import { Firestore } from '@google-cloud/firestore'
-import { HasId, OmitId, Encodable, Decodable } from './types'
+import { HasId, OmitId, AdminEncodable, AdminDecodable } from './types'
 import { Context } from './context'
 import { AdminCollection } from './collection'
 import { AdminQuery } from './query'
@@ -13,8 +13,8 @@ export class FirestoreSimpleAdmin {
 
   collection<T extends HasId, S = OmitId<T>> ({ path, encode, decode }: {
     path: string,
-    encode?: Encodable<T, S>,
-    decode?: Decodable<T, S>,
+    encode?: AdminEncodable<T, S>,
+    decode?: AdminDecodable<T, S>,
   }): AdminCollection<T, S> {
     const factory = new CollectionFactory<T, S>({
       context: this.context,
@@ -25,8 +25,8 @@ export class FirestoreSimpleAdmin {
   }
 
   collectionFactory<T extends HasId, S = OmitId<T>> ({ encode, decode }: {
-    encode?: Encodable<T, S>,
-    decode?: Decodable<T, S>,
+    encode?: AdminEncodable<T, S>,
+    decode?: AdminDecodable<T, S>,
   }): CollectionFactory<T, S> {
     return new CollectionFactory<T, S>({
       context: this.context,
@@ -37,7 +37,7 @@ export class FirestoreSimpleAdmin {
 
   collectionGroup<T extends HasId, S = OmitId<T>> ({ collectionId, decode }: {
     collectionId: string,
-    decode?: Decodable<T, S>,
+    decode?: AdminDecodable<T, S>,
   }): AdminQuery<T, S> {
     const query = this.context.firestore.collectionGroup(collectionId)
     const converter = new AdminConverter({ decode })
@@ -55,13 +55,13 @@ export class FirestoreSimpleAdmin {
 
 class CollectionFactory<T extends HasId, S = OmitId<T>> {
   context: Context
-  encode?: Encodable<T, S>
-  decode?: Decodable<T, S>
+  encode?: AdminEncodable<T, S>
+  decode?: AdminDecodable<T, S>
 
   constructor ({ context, encode, decode }: {
     context: Context,
-    encode?: Encodable<T, S>,
-    decode?: Decodable<T, S>,
+    encode?: AdminEncodable<T, S>,
+    decode?: AdminDecodable<T, S>,
   }) {
     this.context = context
     this.encode = encode
